replace build_multi with zig implementation

+++ build_multi.py
@@ -5,13 +5,14 @@ import json
 import shutil
 import glob
 import time
-import concurrent.futures
+from multiprocessing.pool import ThreadPool
 
 version = "0.6.0"
 out_path = "zig-out/bin/"
 windows = f"{out_path}mufiz.exe"
 bin = "zig-out/bin/mufiz"
 wasm_bin = "zig-out/bin/mufiz.wasm"
+
 
 class PackageBuilder:
     def __init__(self, version, out_path):
@@ -20,8 +21,8 @@ class PackageBuilder:
         self.data = self.load_targets()
 
     def load_targets(self):
-        with open('targets.json', 'r') as file:
-            return json.load(file)['targets']
+        with open("targets.json", "r") as file:
+            return json.load(file)["targets"]
 
     def command_str(self, arch, target, pkg):
         return f"fpm -v {self.version} -a {arch} -s zip -t {pkg} --prefix /usr/bin -m 'Mustafif0929@gmail.com' --description 'The Mufi Programming Language' -n mufiz ./zig-out/bin/mufiz_{self.version}_{target}.zip "
@@ -30,9 +31,15 @@ class PackageBuilder:
         command = self.command_str(arch, target, pkg)
         subprocess.run(command, shell=True, text=True)
         if pkg == "deb":
-            shutil.move(f"mufiz_{self.version}_{arch}.{pkg}", f"mufiz_{self.version}_{target}.{pkg}")
+            shutil.move(
+                f"mufiz_{self.version}_{arch}.{pkg}",
+                f"mufiz_{self.version}_{target}.{pkg}",
+            )
         elif pkg == "rpm":
-            shutil.move(f"mufiz-{self.version}-1.{arch}.{pkg}", f"mufiz_{self.version}_{target}.{pkg}")
+            shutil.move(
+                f"mufiz-{self.version}-1.{arch}.{pkg}",
+                f"mufiz_{self.version}_{target}.{pkg}",
+            )
         print(f"Built {pkg} package for {target}")
 
     def build_target(self, target):
@@ -42,20 +49,21 @@ class PackageBuilder:
         else:
             command = f"zig build -Doptimize=ReleaseSmall -Dtarget={target} -Denable_net=false -Denable_fs=false -Dsandbox=true"
         subprocess.run(command, shell=True, text=True)
-        if "x86_64-windows" in target or "aarch64-windows" in target:
+        if "windows" in target:
             windows_zip = f"mufiz_{self.version}_{target}.zip"
-            with zipfile.ZipFile(self.out_path + windows_zip, 'w') as wz:
+            with zipfile.ZipFile(self.out_path + windows_zip, "w") as wz:
                 wz.write(windows, os.path.basename(windows))
-            os.remove(windows)
+            if os.path.exists(windows):
+                os.remove(windows)
             print(f"Zipped successfully {windows_zip}")
         elif "wasm32-wasi" in target:
             zipper = f"mufiz_{self.version}_{target}.zip"
-            with zipfile.ZipFile(self.out_path + zipper, 'w') as z:
+            with zipfile.ZipFile(self.out_path + zipper, "w") as z:
                 z.write(wasm_bin, os.path.basename(wasm_bin))
             os.remove(wasm_bin)
         else:
             zipper = f"mufiz_{self.version}_{target}.zip"
-            with zipfile.ZipFile(self.out_path + zipper, 'w') as z:
+            with zipfile.ZipFile(self.out_path + zipper, "w") as z:
                 z.write(bin, os.path.basename(bin))
             os.remove(bin)
             print(f"Zipped successfully {zipper}")
@@ -65,7 +73,7 @@ class PackageBuilder:
         arch_map_deb = {
             "x86_64-linux": "amd64",
             "aarch64-linux": "arm64",
-           # "arm-linux": "arm",
+            # "arm-linux": "arm",
             "mips64-linux-musl": "mips64",
             "mips64el-linux-musl": "mips64el",
             "mipsel-linux-musl": "mipsel",
@@ -73,13 +81,13 @@ class PackageBuilder:
             "powerpc64-linux": "powerpc64",
             "powerpc64le-linux": "powerpc64le",
             "powerpc-linux": "powerpc",
-            "riscv64-linux": "riscv64"
+            "riscv64-linux": "riscv64",
         }
-        
+
         arch_map_rpm = {
             "x86_64-linux": "x86_64",
             "aarch64-linux": "aarch64",
-           # "arm-linux": "arm",
+            # "arm-linux": "arm",
             "mips64-linux-musl": "mips64",
             "mips64el-linux-musl": "mips64el",
             "mipsel-linux-musl": "mipsel",
@@ -87,9 +95,9 @@ class PackageBuilder:
             "powerpc64-linux": "ppc64",
             "powerpc64le-linux": "ppc64le",
             "powerpc-linux": "ppc",
-            "riscv64-linux": "riscv64"
+            "riscv64-linux": "riscv64",
         }
-        
+
         if target in (tuple(arch_map_deb.keys())):
             deb_arch = arch_map_deb[target]
             rpm_arch = arch_map_rpm[target]
@@ -97,10 +105,18 @@ class PackageBuilder:
             self.build_package(rpm_arch, target, "rpm")
 
     def build_packages(self):
-        for target in self.data:
-            self.build_target(target)
+        # Create a thread pool with a maximum of 2 threads
+        pool = ThreadPool(2)
 
-        os.remove(self.out_path + "mufiz.pdb")
+        # Use the thread pool to concurrently build two targets at a time
+        pool.map(self.build_target, self.data)
+
+        # Close the thread pool and wait for all threads to complete
+        pool.close()
+        pool.join()
+
+        if os.path.exists(self.out_path + "mufiz.pdb"):
+            os.remove(self.out_path + "mufiz.pdb")
 
         for target in self.data:
             self.build_linux_pkg(target)
